Make XFCE application switcher background opaque
The application switcher in XFCE (the popup that appears when you press
Alt-Tab) has a style which make the background transparent. If you have
a dark background and dark themes in the windows, only the current
selected line is visible as you scroll over it.

This patch adds an opaque background to that window class in particular,
trying to mimic the style as it was in XFCE 4.12 with the Human theme of
Ubuntu 16.04, although I admittedly have not ensured that the colors and
borders are pixel-perfect.